+++ oidc_provider/settings.py
@@ -62,6 +62,13 @@ class DefaultSettings(object):
         return 'oidc_provider.lib.utils.common.default_sub_generator'
 
     @property
+    def OIDC_SESSION_MANAGEMENT_ENABLE(self):
+        """
+        OPTIONAL. If enabled, the Server will support Session Management 1.0 specification.
+        """
+        return False
+
+    @property
     def OIDC_SKIP_CONSENT_ALWAYS(self):
         """
         OPTIONAL. If enabled, the Server will NEVER ask the user for consent.